+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 #v0.1
 
-- Add the Core of Heroes III (HOTA).
-- Add the End Point about Castle's creatures.
+- Add the core of Heroes III (HOTA).
+- Add the end point about Castle's creatures.
+- Add the end point about Rampart's creatures.